+++ src/js/language/da_DK.js
@@ -315,6 +315,7 @@
                 SE: 'svensk',
                 SI: 'slovensk',
                 SK: 'slovakisk',
+                VE: 'venezuelansk',
                 ZA: 'sydafrikansk'
             }
         },