+++ src/js/language/pl_PL.js
@@ -315,6 +315,7 @@
                 SE: 'szwedzki',
                 SI: 'słoweński',
                 SK: 'słowacki',
+                VE: 'wenezuelski',
                 ZA: 'południowo Afrykański'
             }
         },