+++ src/js/language/tr_TR.js
@@ -315,6 +315,7 @@
                 SE: 'İsveç',
                 SI: 'Sloven',
                 SK: 'Slovak',
+                VE: 'Venezüella',
                 ZA: 'Kuzey Afrikalı'
             }
         },